Refactored code and updated hermit deps action
